Move getProposedFacetCuts to deployHelpers
