+++ packages/deployments/contracts/src/index.ts
@@ -5,4 +5,3 @@ export { Deployments } from "./deployments";
 export * from "./abi";
 export * from "./contracts";
 export * from "./abis";
-export * from "./getProposedFacetCuts";
